Discord/requirements.txt
discord.py==1.7.3
psycopg==3.0.8
python-dotenv==0.19.2